move concatTags to TagsHelper
rename method
