+++ internal-api/src/main/java/datadog/trace/util/TagsHelper.java
@@ -66,4 +66,13 @@ public final class TagsHelper {
         || c == '/'
         || c == ':';
   }
+
+  public static String concatTags(String... tags) {
+    StringBuilder sb = new StringBuilder();
+    for (String tag : tags) {
+      sb.append(tag);
+      sb.append(",");
+    }
+    return sb.substring(0, sb.length() - 1); // Remove last comma
+  }
 }
